+++ src/errors/AliasUpdateError.ts
@@ -1,0 +1,6 @@
+export default class AliasUpdateError extends Error {
+  constructor(message = "Failed to update index alias") {
+    super(message);
+    this.name = "AliasUpdateError";
+  }
+}