+++ src/errors/IndexAlreadyExistsError.ts
@@ -1,0 +1,6 @@
+export default class IndexAlreadyExistsError extends Error {
+  constructor(message = "Index already exists") {
+    super(message);
+    this.name = "IndexAlreadyExistsError";
+  }
+}