+++ src/errors/IndexNotFoundError.ts
@@ -1,0 +1,6 @@
+export default class IndexNotFoundError extends Error {
+  constructor(message = "Index does not exist as a concrete index or an alias") {
+    super(message);
+    this.name = "IndexNotFoundError";
+  }
+}