+++ src/errors/ReindexVerificationError.ts
@@ -1,0 +1,6 @@
+export default class ReindexVerificationError extends Error {
+  constructor(message = "Document count mismatch after reindex") {
+    super(message);
+    this.name = "ReindexVerificationError";
+  }
+}